requirements.txt
requests
streamlit
openai
python-dotenv
numpy
